- made ReflectionProbeManager independent of RenderView - moved ReflectionProbeManager and RenderList generation out of RenderView and into Renderer - renamed BufferStackAllocator to LinearGPUBufferAllocator - fixed bug in Physics, causing collision meshes to not be loaded/used

+++ VEngine2/src/component/Level.cpp
@@ -119,7 +119,7 @@ bool Level::load(const char *path) noexcept
 			return false;
 		}
 
-		componentIDMappingTable[i] = index;
+		componentIDMappingTable[i] = static_cast<uint32_t>(index);
 	}
 
 	uint32_t nodeCount = 0;
